refactor: extract RejectCode into ic-response-codes
