+++ ic-response-codes/README.md
@@ -1,0 +1,7 @@
+# ic-response-codes
+
+The codes in the response of Internet Computer API request or inter-canister call.
+
+So far, this crate includes:
+
+* [`RejectCode`] - [Reject codes](https://internetcomputer.org/docs/current/references/ic-interface-spec/#reject-codes)
